refactored out interfaces out of App.vue
