+++ src/App.vue
@@ -2,52 +2,7 @@
 import { RouterLink, RouterView } from 'vue-router';
 import { ref, watch } from 'vue';
 
-interface iMoniker {
-  name: string;
-  title: string;
-  pronouns: string;
-}
-
-interface iBasicInfo {
-  level: number;
-  ancestry: any; //TODO: should be custom interface
-  job: any; //TODO: should be custom interface
-}
-
-interface iChosenStats {
-  bulk: number;
-  agility: number;
-  mind: number;
-  magic: number;
-  ancestryTrait: any; //TODO: should be custom interface
-}
-
-interface iDerivedStats {
-  bulk: number;
-  agility: number;
-  mind: number;
-  magic: number;
-  size: number;
-  grit: number;
-  scope: number;
-  memory: number;
-  saveTarget: number;
-  hp: number;
-  stress: number;
-  mp: number;
-  recoveries: number;
-  speed: number;
-  dodge: number;
-  adef: number;
-}
-
-interface iCharacterData {
-  moniker: iMoniker;
-  basicInfo: iBasicInfo;
-  chosenStats: iChosenStats;
-  derivedStats: iDerivedStats;
-  loadout: any;
-}
+import type iCharacterData from './interfaces/iCharacterData';
 
 const blankCharacterData: iCharacterData = {
   moniker: { name: '', title: '', pronouns: '' },
